Add initial lexer code
This is the beginning of handing writting the parser and lexer code.
I have decided to move to that because I want to have the ability
to support more custom error codes in the language.

Also, currently LALRPOP does not support comments in the code and
to support that, I would need to write my own lexer. Hand writting
the parser will not be much more work at that point.