Add test for if stmt w/o else clause
The implict else clause will return None. I am not sure if I want
to support implict else stmts when we are evaluating an if expr, but
it makes the most sense right now. Until we can seperate if stmts
(else optional) from if exprs (else required), this will have to
do.